==> PART2_simple_embeddings_and_topic_modeling.ipynb <==
# PART 2: Unsupervised learning with topic modeling, word and document embeddings.

We'll use the following packages:  
-Gensim  
-Sklearn  

We'll take preprocessed data from PART 1, which we load here from pickle file.

For full tutorials, see  
https://www.machinelearningplus.com/nlp/topic-modeling-gensim-python  
https://radimrehurek.com/gensim/auto_examples/index.html  


# root folder of data
DATA_ROOT = r'C:\Users\h01928\Documents\GIT_codes\NLP_kickstart_tutorial' + r'\\'

# load our preprocessed data
import pickle
data=pickle.load(open(DATA_ROOT + 'turkuNLP_preprocessed_data.pickle','rb'))
real_labels = [x['label'] for x in data]

First we create new list of curated documents (many different options here)

# We use lemmas and replace numbers and proper nouns, which might improve results.
docs = []
for sample in data:
    tokens = []
    for k,lemma in enumerate(sample['tokens_lemma']):
        if sample['tokens_pos']=='NUM' or lemma.isdigit():
            tokens.append('NUMERO')
        elif sample['tokens_pos']=='PROPN':
            tokens.append('ERISNIMI')
        else:
            tokens.append(lemma)        
    docs.append(tokens)    

We train our own word2vec and fasttext word embeddings, which are the basic type of embeddings. We use dimension 50.

For details, see:  
https://radimrehurek.com/gensim/auto_examples/tutorials/run_word2vec.html  
https://radimrehurek.com/gensim/auto_examples/tutorials/run_fasttext.html  

# https://radimrehurek.com/gensim/auto_examples/tutorials/run_word2vec.html
# https://radimrehurek.com/gensim/auto_examples/tutorials/run_doc2vec_lee.html
import gensim.models

# train word2vec model for tokens
print('Training word2vec')
word2vec_model = gensim.models.Word2Vec(sentences=docs,min_count=5,size=50)
print('Word2vec vector for "raha":\n%s' % str(word2vec_model.wv["raha"]))

fasttext_model = gensim.models.FastText(size=50, window=3, min_count=2)
# build the vocabulary
fasttext_model.build_vocab(docs)
# train the model
print('Training Fasttext')
fasttext_model.train(sentences=docs, epochs=10,total_examples=fasttext_model.corpus_count, total_words=fasttext_model.corpus_total_words)
print('Fasttext vector for "raha":\n%s' % str(fasttext_model.wv["raha"]))

We can also get simple or shallow embeddings for complete documents using doc2vec, which takes all tokens (in a document) and an unique label of document, such as indices 1,2,3,... Again, we use dimension 50.
For complete tutorial, see  
https://radimrehurek.com/gensim/auto_examples/tutorials/run_doc2vec_lee.html

import gensim.models
# initialize the doc2vec model
doc2vec_model = gensim.models.doc2vec.Doc2Vec(vector_size=50, min_count=2, epochs=40,seed=1)
# create tagged corpus
tagged_docs = [gensim.models.doc2vec.TaggedDocument(x, [i]) for i,x in enumerate(docs)]
# build vocabulary
print('Building vocabulary')
doc2vec_model.build_vocab(tagged_docs)
# train the model
print('Training doc2vec')
doc2vec_model.train(tagged_docs,total_examples=doc2vec_model.corpus_count,epochs=doc2vec_model.epochs)

# get vectors for our documents
doc_vectors = [doc2vec_model.infer_vector(x) for x in docs]
# vector of first document
print("vector of the first document\n%s" % str(doc_vectors[0]))

# visualize document embeddings with t-SNE
from sklearn.manifold import TSNE # final reduction
import numpy as np # array handling

def reduce_dimensions(X):
    num_dimensions = 2  # final num dimensions (2D, 3D, etc)
    # convert both lists into numpy vectors for reduction
    vectors = np.asarray(X)
    # reduce using t-SNE
    vectors = np.asarray(vectors)
    tsne = TSNE(n_components=num_dimensions, random_state=5)
    vectors = tsne.fit_transform(vectors)

    x_vals = [v[0] for v in vectors]
    y_vals = [v[1] for v in vectors]
    return np.array(x_vals), np.array(y_vals)

# get 2D embeddings for each document
x_vals, y_vals = reduce_dimensions(doc_vectors)

print('Plotting document vectors')
%matplotlib notebook
from matplotlib import pyplot as plt
plt.figure(figsize=(8,8))
for label in np.unique(real_labels):
    ind=np.where(np.array(real_labels)==label)[0]
    plt.scatter(x_vals[ind], y_vals[ind],s=18,label=label)
plt.legend(loc='lower right')
plt.show()

As you see, documents are nicely separated (even linearly) by their label despite the fact that we're using unsupervised model that has no knowledge of actual labels!
We could use these vectors to build a good classifier.

Finally, we do Latent Diriclet Allocation (LDA) topic modeling. This is a state of art model for traditional (non-neural) topic modeling. Still useful.

Topic modeling can be easily compared to clustering. As in the case of clustering, the number of topics, like the number of clusters, is a hyperparameter. By doing topic modeling we build clusters of words rather than clusters of texts. A text is thus a mixture of all the topics, each having a certain weight.

Wikipedia: "LDA is a generative statistical model that allows sets of observations to be explained by unobserved groups that explain why some parts of the data are similar. For example, if observations are words collected into documents, it posits that each document is a mixture of a small number of topics and that each word's presence is attributable to one of the document's topics. LDA is an example of a topic model."

import copy
docs_expanded = copy.deepcopy(docs) # create a copy

from gensim.models import Phrases
# Add bigrams and trigrams to docs
bigram = Phrases(docs_expanded)
for idx in range(len(docs_expanded)):
    for token in bigram[docs_expanded[idx]]:
        if '_' in token:
            # Token is a bigram, add to document.
            docs_expanded[idx].append(token)

from gensim.corpora import Dictionary

# Create a dictionary representation of the documents.
dictionary = Dictionary(docs_expanded)

# Filter out words that occur less than 2 documents, or more than 50% of the documents
dictionary.filter_extremes(no_below=2, no_above=0.50)

# Bag-of-words representation of the documents.
corpus = [dictionary.doc2bow(doc) for doc in docs_expanded]

print('Number of unique tokens: %d' % len(dictionary))
print('Number of documents: %d' % len(corpus))

from gensim.models import LdaModel

# Make a index to word dictionary.
temp = dictionary[0]  # This is only to "load" the dictionary.
id2word = dictionary.id2token

# now we train the LDA model
print('Training LDA model')
lda_model = LdaModel(corpus=corpus,id2word=id2word,num_topics=2,passes=40,alpha='auto',eta='auto',iterations = 500,random_state=0)

# print top words and weights 
top_words = lda_model.print_topics(num_topics=2,num_words=15)
for topic in top_words:
    print('Topic %i top terms:' % topic[0])
    for k,word in enumerate(topic[1].split(' + ')):
        print('  %i: %s' % (k+1,word))

# get most probable topic for each sample
inferred_labels = [lda_model[x][0][0] for x in corpus]
# attach label (which we know here)
inferred_labels = ['TERVEYS' if x==0 else 'TALOUS' for x in inferred_labels]
# compute accuracy
accuracy = (sum([inferred_labels[k]==real_labels[k] for k in range(len(real_labels))])/len(real_labels))
# in case topics were flipped, take the better one
accuracy = max(accuracy,1-accuracy)

print('\nAccuracy of topic labels against the real labels was %f' % accuracy )

Texts were divided into correct classes with accuracy 94%! Again, this was unsupervised method, but we could have used it to label our texts if haven't known real labels. However, the results depends strongly on choice of parameters in LDA and not all give such nice separation.
